+++ CHANGES.md
@@ -1,11 +1,12 @@
 **Added:**
-* 
+* method in HeaderFile for getting inherited assignments
 
 **Changed:**
-* 
+* made HeaderClass have a reference to its owner file
+    * this required modified the grammar _slightly_
 
 **Fixed:**
-* forgot to include the locale in getting resource bundles
+* 
 
 **Notes:**
 *